use decorators to define routes
