+++ packages/cli/src/ResponseHelper.ts
@@ -146,9 +146,12 @@ const isUniqueConstraintError = (error: Error) =>
  * @returns
  */
 
-export function send(processFunction: (req: Request, res: Response) => Promise<any>, raw = false) {
+export function send<T, R extends Request, S extends Response>(
+	processFunction: (req: R, res: S) => Promise<T>,
+	raw = false,
+) {
 	// eslint-disable-next-line consistent-return
-	return async (req: Request, res: Response) => {
+	return async (req: R, res: S) => {
 		try {
 			const data = await processFunction(req, res);
