+++ packages/cli/src/UserManagement/middlewares/index.ts
@@ -1,0 +1,2 @@
+/* eslint-disable import/no-cycle */
+export * from './auth';
